Factored the Vapor code out of FOSMVVM into FOSMVVMVapor
