+++ Sources/FOSMVVM/Protocols/ViewModelRequest.swift
@@ -63,40 +63,3 @@ public extension ViewModelRequest {
         responseBody ?? .stub()
     }
 }
-
-#if canImport(Vapor)
-import Vapor
-
-public extension RoutesBuilder {
-    // NOTE: At this time, an overload that registers a
-    //   ViewModelRequest directly is not provided.  The
-    //   thinking is that by keeping ViewModel and its
-    //   ViewModelRequest tightly bound vai RequestableViewModel,
-    //   the entire suite of types corresponding to a ViewModel
-    //   is type bound and thus checked at compile time.  This
-    //   directs the user to the types that they need to implement.
-
-    /// Enables a ``ViewModel`` instance to be available via
-    /// Vapor's [routing](https://docs.vapor.codes/basics/routing/)
-    ///
-    /// ## Example
-    ///
-    /// ```swift
-    /// func routes(_ app: Application) throws {
-    ///    try app.routes.register(model: LandingPageViewModel.self)
-    /// }
-    /// ```
-    ///
-    /// > Note: ``ServerRequest`` provides a protocol extension that
-    /// >   sets ``ServerRequest/path`` based on the ``ServerRequest/RequestBody``
-    /// >   and ``ServerRequest/ResponseBody`` types.   Thus, the route's path is
-    /// >   automatically maintained and there is never a path collision or confusion between
-    /// >   the client and server.
-    ///
-    /// - Parameter viewModel: A ``RequestableViewModel`` to register
-    func register<Model>(viewModel: Model.Type) throws where Model: RequestableViewModel & VaporViewModelFactory, Model.Request.ResponseBody == Model {
-        try register(collection: VaporServerRequestHost<Model.Request>())
-    }
-}
-
-#endif
