+++ Tests/FOSMVVMTests/.VersionedTestJSON/TestVersionedViewModel_5.6.7.json
@@ -1,0 +1,1 @@
+{"p1":{"wrappedValue":{"vmId":{"ts":1738194098.507102}},"vFirst":"1.0.0","vLast":"1.0.0"},"aLocalizedString":"","p3":{"vFirst":"1.0.0","wrappedValue":{"vmId":{"ts":1738194098.5071182}},"vLast":"2.1.0"},"p4":{"wrappedValue":{"vmId":{"ts":1738194098.5071213}},"vFirst":"1.0.0","vLast":"3.0.0"},"vmId":{"ts":1738194098.507123},"p2":{"vFirst":"1.0.0","wrappedValue":{"vmId":{"ts":1738194098.5071135}},"vLast":"2.0.0"}}
